refactor `bindAddressFormEvents`
this now accepts an `options` object, with which you can specify 
different attributes of the form

`fields`: an array of objects that correspond to the different inputs 
that should be bound. each field object has a `selector`, a `message` 
(used in error handling, and an optional `validation`. by default every 
field is validated for the presence of `value`, but something like a 
checkbox needs its `checked` attribute... checked. the form is hard 
wired to check for an `address` field, a `legal` field, and an `email` 
field and run separate logic for each if included. these fields are 
identified by the `name` key.
`multiples`: a selector for where to find an element to hold multiple 
responses for a given address
`errors`: selector to an element to render validation errors
`form`: selector to the form element to which the listener is attached